moved tests to other test suite
